fix(wiki): add /onboarding route (closes #213)
Plan 26 §C6 expects GET /onboarding to be reachable but the wiki app only
exposed the wizard inline at the gated home route '/'. Extract the wizard
into a reusable <OnboardingWizard /> component and mount it at
/(public)/onboarding/page.tsx so the route is deep-linkable. Authenticated
users always see the wizard there (no redirect-on-onboardedAt) so they can
revisit/replay; unauthenticated users still bounce to /login. The home
route '/' keeps its existing auth + onboarded-redirect behavior, now
delegating render to the same wizard component.
